MainPage in ErsterDurchlauf kopiert, MainPage Inhalt gelöscht.
