+++ Spiel/Spiel/Spiel/MainPage.xaml
@@ -6,45 +6,7 @@
 
     <ContentPage.Content>
         <AbsoluteLayout>
-            <Label x:Name="Label1" Text="1" FontSize="78" TextColor="Black" AbsoluteLayout.LayoutBounds="0, 0, 120, 120"
-                  AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0" />
-            <Label x:Name="Label2" Text="2" FontSize="78" XAlign="End"  YAlign="Start" TextColor="Black" AbsoluteLayout.LayoutBounds="1, 0, 120, 120"
-                  AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0"/>
-            <Label x:Name="Label3" Text="3" XAlign="Start" YAlign="End" FontSize="78" TextColor="Black" AbsoluteLayout.LayoutBounds="0, 1, 120, 120"
-                  AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0"/>
-            <Label x:Name="Label4" Text="4" FontSize="78" TextColor="Black" XAlign="End" YAlign="End" AbsoluteLayout.LayoutBounds="1, 1, 120, 120"
-                  AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0"/>
-            <BoxView x:Name="Box1" Color="Green"  AbsoluteLayout.LayoutBounds="0, 0, 40, 40"
-                AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0"/>
-            <BoxView x:Name="Box2" Color="Green"  AbsoluteLayout.LayoutBounds="1, 0, 40, 40"
-                AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0"/>
-            <BoxView x:Name="Box3" Color="Green"  AbsoluteLayout.LayoutBounds="0, 1, 40, 40"
-                AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0" />
-            <BoxView x:Name="Box4" Color="Green"  AbsoluteLayout.LayoutBounds="1, 1, 40, 40"
-                AbsoluteLayout.LayoutFlags="PositionProportional" Opacity="0" />
-            <Label x:Name="Countdown10" Text="10" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="1"/>
-            <Label x:Name="Countdown9" Text="9" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown8" Text="8" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown7" Text="7" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown6" Text="6" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown5" Text="5" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown4" Text="4" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown3" Text="3" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown2" Text="2" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Countdown1" Text="1" XAlign="Center" YAlign="Center" FontSize="66" TextColor="DarkBlue" AbsoluteLayout.LayoutBounds="0.5, 0.5, 100, 100" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0"/>
-            <Label x:Name="Fragezeichen" Text="?" XAlign="Center" YAlign="Center" FontSize="72" TextColor="Green" AbsoluteLayout.LayoutBounds="0.5, 0.5, 120, 120" 
-                   AbsoluteLayout.LayoutFlags="All" Opacity="0" />
-
+            <Label Text="Test" />
         </AbsoluteLayout>
     </ContentPage.Content>
